requirements.txt
streamlit==1.28.0
pandas==2.1.0
numpy==1.24.0
plotly==5.17.0
pinecone-client==3.0.0
requests==2.31.0
python-dotenv==1.0.0
openpyxl==3.1.2
scikit-learn==1.3.0
altair==5.1.0
reportlab==4.0.4
